Added a Redux Store
Create a store (suing a reducer)
Create a reducer => any function that works on actions
dispatch => invoke the action using action type
Subscribe to store. this should be done before dispatch calls.
ComponentDidmount is best place to dfine the subscribe method.

Now, the ajax call result is set in the store
and component is rendered using the store.getState().properties